Rename directories rtx->rtx4 rtx2->rtx5
